Rename src/gatsby/ to src/gatsby-node/
